+++ src/$createImageNode.ts
@@ -2,6 +2,6 @@ import { $applyNodeReplacement } from 'lexical';
 import ImageNode from './ImageNode';
 import ImageInfo from './private/ImageInfo';
 
-export function $createImageNode({ alt, height, src, width }: ImageInfo): ImageNode {
+export default function $createImageNode({ alt, height, src, width }: ImageInfo): ImageNode {
   return $applyNodeReplacement(new ImageNode(src, width, height, alt));
 }