+++ src/ImageNode.tsx
@@ -10,7 +10,7 @@ import {
   Spread,
 } from 'lexical';
 import { JSX } from 'react';
-import { $createImageNode } from './$createImageNode';
+import $createImageNode from './$createImageNode';
 import ImageInfo from './private/ImageInfo';
 
 function $convertImageElement(domNode: Node): null | DOMConversionOutput {