+++ src/index.tsx
@@ -1,3 +1,4 @@
+export { default as $createImageNode } from './$createImageNode';
 export { default as $isImageNode } from './$isImageNode';
 export { default as ImageNode } from './ImageNode';
 export { default as ImagePlugin } from './ImagePlugin';