+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## Unreleased
+
+- Added `$createImageNode` function export
+- Added `$isImageNode` function export
+
 ## 0.1.0 - 2025-05-16
 
 - Added `ImageNode`